Start to work with shaders -- vertex shader and fragment shader. Store info into vao->vbo->ebo.

+++ mario/src/main/java/jade/LevelScene.java
@@ -9,7 +9,7 @@ public class LevelScene extends Scene{
   }
 
   @Override
-  public void update(float dt) {
+  public void update(double dt) {
 
   }
