+++ mario/src/main/java/util/Time.java
@@ -3,11 +3,9 @@ package util;
 public class Time {
   public static long timeStarted = System.nanoTime();
 
-  public static float getTime() {
+  public static double getTime() {
     // The time is in nanosecond, we need to convert to second.
-    return (float) ((System.nanoTime() - timeStarted) * 1E-9); //conversion
+    return ((System.nanoTime() - timeStarted) * 1E-9); //conversion
   }
-
-
 
 }
